LayoutTests/fast/js/const-expected.txt

This test checks that const declarations in JavaScript work and are readonly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS x is "RIGHT"
PASS y is "RIGHT"
PASS successfullyParsed is true

TEST COMPLETE
